+++ src/mizdb_tomselect/widgets.py
@@ -16,9 +16,9 @@ class MIZSelect(forms.Select):
         self,
         model,
         url="autocomplete",
-        search_lookup="name__icontains",
-        value_field="id",
-        label_field="name",
+        value_field="",
+        label_field="",
+        search_lookup="",
         create_field="",
         multiple=False,
         changelist_url="",
@@ -33,12 +33,15 @@ class MIZSelect(forms.Select):
             model: the django model that the choices are derived from
             url: the URL pattern name of the view that serves the choices and
               handles requests from the TomSelect element
+            value_field: the name of the model field that corresponds to the
+              choice value of an option (f.ex. 'id'). Defaults to the name of
+              the model's primary key field.
+            label_field: the name of the model field that corresponds to the
+              human-readable value of an option (f.ex. 'name'). Defaults to the
+              value of the model's `name_field` attribute. If the model has no
+              `name_field` attribute, it defaults to 'name'.
             search_lookup: a Django field lookup to use with the given search
               term to filter the results
-            value_field: the name of the model field that corresponds to the
-              choice value of an option (f.ex. 'id')
-            label_field: the name of the model field that corresponds to the
-              human-readable value of an option (f.ex. 'name')
             create_field: the name of the model field used to create new
               model objects with
             multiple: if True, allow selecting multiple options
@@ -52,10 +55,10 @@ class MIZSelect(forms.Select):
         """
         self.model = model
         self.url = url
-        self.search_lookup = search_lookup
+        self.value_field = value_field or self.model._meta.pk.name
+        self.label_field = label_field or getattr(self.model, "name_field", "name")
+        self.search_lookup = search_lookup or f"{self.label_field}__icontains"
         self.create_field = create_field
-        self.value_field = value_field
-        self.label_field = label_field
         self.multiple = multiple
         self.changelist_url = changelist_url
         self.add_url = add_url
@@ -110,7 +113,7 @@ class MIZSelect(forms.Select):
 class MIZSelectTabular(MIZSelect):
     """A MIZSelect widget that displays results in a table with a table header."""
 
-    def __init__(self, *args, extra_columns=None, value_field_label="ID", label_field_label="Object", **kwargs):
+    def __init__(self, *args, extra_columns=None, value_field_label="", label_field_label="", **kwargs):
         """
         Instantiate a MIZSelectTabular widget.
 
@@ -119,14 +122,16 @@ class MIZSelectTabular(MIZSelect):
               for additional columns. The field name tells TomSelect what
               values to look up on a model object result for a given column.
               The label is the table header label for a given column.
-            value_field_label: table header label for the value field column
-            label_field_label: table header label for the label field column
+            value_field_label: table header label for the value field column.
+              Defaults to value_field.title().
+            label_field_label: table header label for the label field column.
+              Defaults to the verbose_name of the model.
             args: additional positional arguments passed to MIZSelect
             kwargs: additional keyword arguments passed to MIZSelect
         """
         super().__init__(*args, **kwargs)
-        self.value_field_label = value_field_label
-        self.label_field_label = label_field_label
+        self.value_field_label = value_field_label or self.value_field.title()
+        self.label_field_label = label_field_label or self.model._meta.verbose_name or "Object"
         self.extra_columns = extra_columns or {}
 
     def build_attrs(self, base_attrs, extra_attrs=None):